week8/requirements.txt
langchain
langchain-community
langchain-ollama
faiss-cpu
pydantic
python-dotenv
langsmith
tiktoken
